♻️ Refactor : [임시] 관련로직 제외 주석처리
