+++ module-ticket/src/main/java/com/example/moduleticket/ModuleTicketApplication.java
@@ -2,6 +2,7 @@ package com.example.moduleticket;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
 public class ModuleTicketApplication {
