+++ module-ticket/src/main/java/com/example/moduleticket/config/OpenFeignConfig.java
@@ -1,0 +1,10 @@
+package com.example.moduleticket.config;
+
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableFeignClients("com.example.moduleticket.feign")
+public class OpenFeignConfig {
+
+}
