+++ module-ticket/src/main/java/com/example/moduleticket/domain/auction/repository/AuctionRepositoryQuery.java
@@ -1,10 +1,10 @@
-package com.example.moduleticket.domain.auction.repository;
-
-import com.example.moduleticket.domain.auction.dto.request.AuctionSearchCondition;
-import com.example.moduleticket.domain.auction.entity.Auction;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-public interface AuctionRepositoryQuery {
-	Page<Auction> findByConditions(AuctionSearchCondition dto, Pageable pageable);
-}
+//package com.example.moduleticket.domain.auction.repository;
+//
+//import com.example.moduleticket.domain.auction.dto.request.AuctionSearchCondition;
+//import com.example.moduleticket.domain.auction.entity.Auction;
+//import org.springframework.data.domain.Page;
+//import org.springframework.data.domain.Pageable;
+//
+//public interface AuctionRepositoryQuery {
+//	Page<Auction> findByConditions(AuctionSearchCondition dto, Pageable pageable);
+//}
