+++ module-ticket/src/main/java/com/example/moduleticket/domain/auction/repository/AuctionTicketInfoRepositoryQuery.java
@@ -1,9 +1,9 @@
-package com.example.moduleticket.domain.auction.repository;
-
-
-import com.example.moduleticket.domain.auction.dto.AuctionTicketInfoDto;
-import com.example.moduleticket.domain.ticket.entity.Ticket;
-
-public interface AuctionTicketInfoRepositoryQuery {
-	AuctionTicketInfoDto findTicketInfo(Ticket ticket);
-}
+//package com.example.moduleticket.domain.auction.repository;
+//
+//
+//import com.example.moduleticket.domain.auction.dto.AuctionTicketInfoDto;
+//import com.example.moduleticket.domain.ticket.entity.Ticket;
+//
+//public interface AuctionTicketInfoRepositoryQuery {
+//	AuctionTicketInfoDto findTicketInfo(Ticket ticket);
+//}
