--- module-ticket/src/main/java/com/example/moduleticket/domain/config/IamPortConfig.java
@@ -1,21 +1,0 @@
-package com.example.moduleticket.domain.config;
-
-import com.siot.IamportRestClient.IamportClient;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class IamPortConfig {
-	
-	@Value("${import.api.key}")
-	private String apiKey;
-	
-	@Value("${import.api.secret}")
-	private String apiSecret;
-	
-	@Bean
-	public IamportClient iamportClient() {
-		return new IamportClient(apiKey, apiSecret);
-	}
-}
