--- module-ticket/src/main/java/com/example/moduleticket/domain/config/JpaConfig.java
@@ -1,17 +1,0 @@
-package com.example.moduleticket.domain.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class JpaConfig {
-	
-	@PersistenceContext
-	private EntityManager entityManager;
-	
-	@Bean
-	public JPAQueryFactory jpaQueryFactory() { return new JPAQueryFactory(entityManager); }
-}
