+++ module-ticket/src/main/java/com/example/moduleticket/domain/point/dto/response/PointExchangeResponse.java
@@ -1,15 +1,15 @@
-package com.example.moduleticket.domain.point.dto.response;
-
-import com.example.moduleticket.domain.point.enums.PointHistoryType;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class PointExchangeResponse {
-
-	private final Long memberId;
-	private final Integer charge;
-	private final Integer point;
-	private final PointHistoryType type;
-}
+//package com.example.moduleticket.domain.point.dto.response;
+//
+//import com.example.moduleticket.domain.point.enums.PointHistoryType;
+//import lombok.Getter;
+//import lombok.RequiredArgsConstructor;
+//
+//@RequiredArgsConstructor
+//@Getter
+//public class PointExchangeResponse {
+//
+//	private final Long memberId;
+//	private final Integer charge;
+//	private final Integer point;
+//	private final PointHistoryType type;
+//}
