+++ module-ticket/src/main/java/com/example/moduleticket/domain/point/entity/PointPayment.java
@@ -1,36 +1,36 @@
-package com.example.moduleticket.domain.point.entity;
-
-import com.example.moduleticket.domain.member.entity.Member;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PointPayment {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-	
-	private Integer price;
-	
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "member_id", nullable = false)
-	private Member member;
-	
-	@Builder
-	public PointPayment(Integer price, Member member) {
-		this.price = price;
-		this.member = member;
-	}
-}
+//package com.example.moduleticket.domain.point.entity;
+//
+//import com.example.moduleticket.domain.member.entity.Member;
+//import jakarta.persistence.Entity;
+//import jakarta.persistence.FetchType;
+//import jakarta.persistence.GeneratedValue;
+//import jakarta.persistence.GenerationType;
+//import jakarta.persistence.Id;
+//import jakarta.persistence.JoinColumn;
+//import jakarta.persistence.ManyToOne;
+//import lombok.AccessLevel;
+//import lombok.Builder;
+//import lombok.Getter;
+//import lombok.NoArgsConstructor;
+//
+//@Getter
+//@Entity
+//@NoArgsConstructor(access = AccessLevel.PROTECTED)
+//public class PointPayment {
+//
+//	@Id
+//	@GeneratedValue(strategy = GenerationType.IDENTITY)
+//	private Long id;
+//
+//	private Integer price;
+//
+//	@ManyToOne(fetch = FetchType.LAZY)
+//	@JoinColumn(name = "member_id", nullable = false)
+//	private Member member;
+//
+//	@Builder
+//	public PointPayment(Integer price, Member member) {
+//		this.price = price;
+//		this.member = member;
+//	}
+//}
