+++ module-ticket/src/main/java/com/example/moduleticket/domain/point/repository/PointPaymentRepository.java
@@ -1,8 +1,8 @@
-package com.example.moduleticket.domain.point.repository;
-
-
-import com.example.moduleticket.domain.point.entity.PointPayment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PointPaymentRepository extends JpaRepository<PointPayment, Long> {
-}
+//package com.example.moduleticket.domain.point.repository;
+//
+//
+//import com.example.moduleticket.domain.point.entity.PointPayment;
+//import org.springframework.data.jpa.repository.JpaRepository;
+//
+//public interface PointPaymentRepository extends JpaRepository<PointPayment, Long> {
+//}
