+++ module-ticket/src/main/java/com/example/moduleticket/domain/ticket/service/TicketService.java
@@ -6,8 +6,10 @@ import static com.example.modulecommon.exception.ErrorCode.USER_ACCESS_DENIED;
 import com.example.modulecommon.entity.AuthUser;
 import com.example.modulecommon.exception.ServerException;
 import com.example.moduleticket.domain.reservation.dto.ReservationDto;
-import com.example.moduleticket.domain.ticket.dto.GameDto;
-import com.example.moduleticket.domain.ticket.dto.SeatDto;
+import com.example.moduleticket.feign.GameService;
+import com.example.moduleticket.feign.SeatService;
+import com.example.moduleticket.feign.dto.GameDto;
+import com.example.moduleticket.feign.dto.SeatDto;
 import com.example.moduleticket.domain.ticket.dto.TicketContext;
 import com.example.moduleticket.domain.ticket.dto.request.TicketCreateRequest;
 import com.example.moduleticket.domain.ticket.dto.response.TicketResponse;
@@ -35,18 +37,18 @@ public class TicketService {
 	private final TicketRepository ticketRepository;
 	private final TicketSeatService ticketSeatService;
 	private final TicketPaymentService ticketPaymentService;
-	//private final PointService pointService;
 	private final TicketCreateService ticketCreateService;
 	private final SeatHoldRedisUtil seatHoldRedisUtil;
 	private final ApplicationEventPublisher eventPublisher;
+	private final SeatService seatService;
+	private final GameService gameService;
 
 	@Transactional(readOnly = true)
 	public TicketResponse getTicket(AuthUser auth, Long ticketId) {
 		Ticket ticket = ticketRepository.findByIdAndMemberIdAndDeletedAtIsNull(ticketId, auth.getMemberId())
 				.orElseThrow(() -> new ServerException(TICKET_NOT_FOUND));
 		Long gameId = ticket.getGameId();
-		//todo : 게임정보 가져오기
-		GameDto gameDto = new GameDto();
+		GameDto gameDto = gameService.getGame(gameId);
 
 		return convertTicketResponse(ticket, gameDto);
 	}
@@ -56,8 +58,7 @@ public class TicketService {
 		List<Ticket> allTickets = ticketRepository.findAllByMemberIdWithGame(auth.getMemberId());
 
 		List<Long> gameIds = allTickets.stream().map(Ticket::getGameId).toList();
-		//todo : 게임정보 가져오기 bulk select
-		List<GameDto> gameDtos = new ArrayList<>();
+		List<GameDto> gameDtos = gameService.getGames(gameIds);
 		Map<Long, GameDto> gameDtoMap = gameDtos.stream().collect(Collectors.toMap(GameDto::getId, Function.identity()));
 
 		return allTickets.stream().map(ticket -> convertTicketResponse(ticket, gameDtoMap.get(ticket.getGameId()))).toList();
