+++ module-ticket/src/main/resources/application.properties
@@ -1,2 +1,27 @@
 spring.application.name=module-ticket
 server.port=8082
+
+spring.datasource.url=${DB_URL}
+spring.datasource.username=${DB_USERNAME}
+spring.datasource.password=${DB_PASSWORD}
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.use_sql_comments=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
+
+jwt.secret.key=${JWT_SECRET_KEY}
+jwt.access.token=${JWT_ACCESS_TOKEN}
+
+import.api.key=${IMPORT_API_KEY}
+import.api.secret=${IMPORT_API_SECRET}
+
+# AWS S3
+spring.cloud.aws.s3.bucketName=${BUCKET_NAME}
+spring.cloud.aws.region.static=${REGION_VALUE}
+spring.cloud.aws.credentials.access-key=${ACCESS_KEY}
+spring.cloud.aws.credentials.secret-key=${SECRET_KEY}
+
+management.endpoints.web.exposure.include=prometheus
